Revert "Merge branch 'main' into tornjak-dashboard-ui-part2"
This reverts commit 5195029b8ae3e2d08c59d6886b2840738ee34e57, reversing
changes made to e5bf4fc81239c14eb515a6b275a9d166038de6cb.
