+++ SPIRE_BUILD_VERSIONS
@@ -1,8 +1,8 @@
-0.11.0
-0.11.1
-0.11.2
-0.11.3
-0.12.0
-0.12.1
-1.0.0
 1.0.1
+1.0.0
+0.12.1
+0.12.0
+0.11.3
+0.11.2
+0.11.1
+0.11.0
