+++ tornjak-frontend/src/components/entry-create.js
@@ -83,6 +83,8 @@ class CreateEntry extends Component {
         this.TornjakApi.populateEntriesUpdate(this.props.globalServerSelected, this.props.entriesListUpdateFunc, this.props.tornjakMessageFunc)
         this.TornjakApi.refreshSelectorsState(this.props.globalServerSelected, this.props.agentworkloadSelectorInfoFunc);
         this.setState({ selectedServer: this.props.globalServerSelected });
+        this.prepareParentIdAgentsList();
+        this.prepareSelectorsList();
       }
     } else {
       // agent doesnt need to do anything
@@ -99,41 +101,23 @@ class CreateEntry extends Component {
       if (prevProps.globalServerSelected !== this.props.globalServerSelected) {
         this.setState({ selectedServer: this.props.globalServerSelected });
       }
-
       if (prevProps.globalServerInfo !== this.props.globalServerInfo) {
-        if (this.props.globalAgentsList !== undefined && this.props.globalEntriesList !== undefined) {
-          this.prepareParentIdAgentsList();
-        }
+        this.prepareParentIdAgentsList();
         this.prepareSelectorsList();
       }
-
-      if (prevProps.globalAgentsList !== this.props.globalAgentsList || prevProps.globalEntriesList !== this.props.globalEntriesList) {
-        if (this.props.globalAgentsList !== undefined && this.props.globalEntriesList !== undefined) {
-          this.prepareParentIdAgentsList();
-        }
-        this.prepareSelectorsList();
+      if (prevProps.globalAgentsList !== this.props.globalAgentsList) {
+        this.prepareParentIdAgentsList();
       }
-
       if (prevState.parentId !== this.state.parentId) {
         this.prepareSelectorsList();
       }
     } else {
       if (prevProps.globalServerInfo !== this.props.globalServerInfo) {
-        if (this.props.globalAgentsList !== undefined && this.props.globalEntriesList !== undefined) {
-          this.prepareParentIdAgentsList();
-        }
+        this.prepareParentIdAgentsList();
         this.prepareSelectorsList();
       }
-
-      if (prevProps.globalAgentsList !== this.props.globalAgentsList || prevProps.globalEntriesList !== this.props.globalEntriesList) {
-        if (this.props.globalAgentsList !== undefined && this.props.globalEntriesList !== undefined) {
-          this.prepareParentIdAgentsList();
-        }
-        this.prepareSelectorsList();
-      }
-
       if (prevState.parentId !== this.state.parentId)  {
-          this.prepareSelectorsList();
+        this.prepareSelectorsList();
       }
     }
   }
