+++ tornjak-frontend/src/tables/entries-list-table.js
@@ -1,25 +1,38 @@
 import React from "react";
 import { connect } from 'react-redux';
+import { DataTable } from "carbon-components-react";
+import ResetIcon from "@carbon/icons-react/es/reset--alt/20";
 import GetApiServerUri from 'components/helpers';
 import IsManager from 'components/is_manager';
 import axios from 'axios'
 import {
     entriesListUpdateFunc
 } from 'redux/actions';
-import Table from './list-table';
+const {
+    TableContainer,
+    Table,
+    TableHead,
+    TableRow,
+    TableBody,
+    TableCell,
+    TableHeader,
+    TableSelectRow,
+    TableSelectAll,
+    TableToolbar,
+    TableToolbarSearch,
+    TableToolbarContent,
+    TableBatchActions,
+    TableBatchAction,
+} = DataTable;
 
-// EntriesListTable takes in 
-// listTableData: entries data to be rendered on table
-// returns entries data inside a carbon component table with specified functions
-class EntriesListTable extends React.Component {
+class DataTableRender extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
             listData: props.data,
-            listTableData: [{ "id": "0" }],
+            listTableData: [{"id":"0"}],
         };
         this.prepareTableData = this.prepareTableData.bind(this);
-        this.deleteEntry = this.deleteEntry.bind(this);
     }
 
     componentDidMount() {
@@ -112,13 +125,74 @@ class EntriesListTable extends React.Component {
             },
         ];
         return (
-            <div>
-                <Table
-                    listTableData={listTableData}
-                    headerData={headerData}
-                    deleteEntity={this.deleteEntry}
-                />
-            </div>
+            <DataTable
+                isSortable
+                rows={listTableData}
+                headers={headerData}
+                render={({
+                    rows,
+                    headers,
+                    getHeaderProps,
+                    getSelectionProps,
+                    onInputChange,
+                    getPaginationProps,
+                    getBatchActionProps,
+                    getTableContainerProps,
+                    selectedRows,
+                }) => (
+                    <TableContainer
+                        {...getTableContainerProps()}
+                    >
+                        <TableToolbar>
+                            <TableToolbarContent>
+                                <TableToolbarSearch onChange={(e) => onInputChange(e)} />
+                            </TableToolbarContent>
+                            <TableBatchActions
+                                {...getBatchActionProps()}
+                            >
+                                <TableBatchAction
+                                    renderIcon={ResetIcon}
+                                    iconDescription="Delete"
+                                    onClick={() => {
+                                        this.deleteEntry(selectedRows);
+                                    }}
+                                >
+                                    Delete
+                                </TableBatchAction>
+                            </TableBatchActions>
+                        </TableToolbar>
+                        <Table size="short" useZebraStyles>
+                            <TableHead>
+                                <TableRow>
+                                    <TableSelectAll {...getSelectionProps()} />
+                                    {headers.map((header) => (
+                                        <TableHeader {...getHeaderProps({ header })}>
+                                            {header.header}
+                                        </TableHeader>
+                                    ))}
+                                </TableRow>
+                            </TableHead>
+                            <TableBody>
+                                {rows.map((row, key) => (
+                                    <TableRow key={key}>
+                                        <TableSelectRow {...getSelectionProps({ row })} />
+                                        {row.cells.map((cell) => (
+                                            <TableCell key={cell.id}>
+                                                {cell.info.header === "info" ? (
+                                                    <div style={{ overflowX: 'auto', width: "400px" }}>
+                                                        <pre>{cell.value}</pre>
+                                                    </div>
+                                                ) : (
+                                                    cell.value)}
+                                            </TableCell>
+                                        ))}
+                                    </TableRow>
+                                ))}
+                            </TableBody>
+                        </Table>
+                    </TableContainer>
+                )}
+            />
         );
     }
 }
@@ -131,4 +205,4 @@ const mapStateToProps = (state) => ({
 export default connect(
     mapStateToProps,
     { entriesListUpdateFunc }
-)(EntriesListTable)
+)(DataTableRender)
